Nullable (#365)
* Shared lib updated with Nullable context. Partially fixes issue #105.

* Server updated with Nullable context. Partially fixes issue #105.

* Razor projects updated with Nullable context. Partially fixes issue #105.

* 01 updated with Nullable context, including instructions (md). Partially fixes issue #105.

* Updated 02. Dependencies to use Nullable

* Updated code to include nullable considerations.

* Updated instructions to align with nullable.

* Updated dependencies to use nullable context

* Fixes issue #358 and Fixes issue #238

* Sync with previous lesson

* Removed guard on GetUserId to allow workshop to flow without authorization.

* Sync changes across lessons

* Updated 03 with nullability context

* Updated instructions with nullable code

* Cleanup whitespace

* Sync dependencies 04 with previous steps

* Fixed: Starting point 04's server code was actually from a later step 06 or above.

* Sync 04 client changes with previous steps

* Updated client with nullable context enabled

* Sync dependencies 05 with previous changes

* Fixed step 05 controller is from step 06

* Sync with client with previous step

* Updated 05 with nullable context enabled

* Sync with previous steps

* Updated 06 with nullable context enabled.

* Synced dependencies with previous lesson

* Sync client changes with previous step

* Updated 07 with nullable context enabled

* Synced dependencies with previous step

* Synced client with previous lesson

* Updated 08 with nullable conetext enabled

* Sync with prevous lesson

* Updated 09 with nullable context enabled

* Updated src with nullable context enabled

* Updated docs

* Notification interop doesn't work with required.

* Extra sln file

* Fixed link that moved for getting started

* Removed extra comments.

* Clairified the message about EditorRequired.

* Trimmed whitespace

* Supressed nullable warning on change event

* Added EditorRequired to step

* Refactored GetUserId. This code was duplicated in multiple places.

* Updated `!=` to `is not` null

* Fixes: SignoutSessionStateManager obsolete warning. See: https://learn.microsoft.com/en-us/dotnet/core/compatibility/aspnet-core/7.0/wasm-app-authentication

* Added Nullable to orderWithStatus

* Updated Razor Class Library verbage, it was reminicent of .NET Core 3.0/3.1

* Changed `!=` to `is not null`

* The app was updated to include a Minimal API implementation of NotificationsController. However the code in save-point/09 and instructions were not updated. This fixes the issue.

* Pulling up some missed changes from the final exercise

* Removed unused NavigationManager references

* Updated code on Map component

+++ save-points/09-progressive-web-app/BlazingPizza.Server/PizzaApiExtensions.cs
@@ -19,6 +19,10 @@ public static class PizzaApiExtensions
                 // We're storing at most one subscription per user, so delete old ones.
                 // Alternatively, you could let the user register multiple subscriptions from different browsers/devices.
                 var userId = GetUserId(context);
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 var oldSubscriptions = db.NotificationSubscriptions.Where(e => e.UserId == userId);
                 db.NotificationSubscriptions.RemoveRange(oldSubscriptions);
 
@@ -49,9 +53,6 @@ public static class PizzaApiExtensions
 
     }
 
-    private static string GetUserId(HttpContext context)
-    {
-        return context.User.FindFirstValue(ClaimTypes.NameIdentifier);
-    }
+    public static string? GetUserId(HttpContext context) => context.User.FindFirstValue(ClaimTypes.NameIdentifier);
 
 }
